♻️refactor: optimizes dependency injection in controllers.
