+++ backend/src/PeopleHub.Api/Controllers/PersonController.cs
@@ -4,7 +4,7 @@ using Microsoft.AspNetCore.Mvc;
 using PeopleHub.Application.Dtos.IndividualPerson;
 using PeopleHub.Application.Dtos.LegalPerson;
 using PeopleHub.Application.Dtos.Person;
-using PeopleHub.Application.Interfaces.Person;
+using PeopleHub.Application.Routers;
 
 namespace PeopleHub.Api.Controllers;
 
@@ -14,21 +14,18 @@ namespace PeopleHub.Api.Controllers;
 [Authorize]
 public class PersonController : ControllerBase
 {
-    private readonly IPersonService _personService;
+    private readonly IPersonRouter _personRouter;
 
-    public PersonController(IPersonService personService)
+    public PersonController(IPersonRouter personService)
     {
-        _personService = personService;
+        _personRouter = personService;
     }
 
     [HttpPost("individual")]
     [Authorize]
     public async Task<IActionResult> RegisterIndividual([FromBody] RegisterIndividualPersonRequestDto request)
     {
-        var response = await _personService.RegisterIndividualAsync(request);
-
-        if (!response.IsSuccess)
-            return StatusCode(response.StatusCode, response);
+        var response = await _personRouter.RegisterIndividualAsync(request);
 
         return StatusCode(response.StatusCode, response);
     }
@@ -37,10 +34,7 @@ public class PersonController : ControllerBase
     [Authorize]
     public async Task<IActionResult> UpdateIndividual([FromBody] UpdateIndividualPersonRequestDto request)
     {
-        var response = await _personService.UpdateIndividualAsync(request);
-
-        if (!response.IsSuccess)
-            return StatusCode(response.StatusCode, response);
+        var response = await _personRouter.UpdateIndividualAsync(request);
 
         return StatusCode(response.StatusCode, response);
     }
@@ -49,10 +43,7 @@ public class PersonController : ControllerBase
     [Authorize]
     public async Task<IActionResult> DeleteIndividual([FromBody] DeleteIndividualPersonDto request)
     {
-        var response = await _personService.DeleteIndividualAsync(request);
-
-        if (!response.IsSuccess)
-            return StatusCode(response.StatusCode, response);
+        var response = await _personRouter.DeleteIndividualAsync(request);
 
         return StatusCode(response.StatusCode, response);
     }
@@ -61,10 +52,7 @@ public class PersonController : ControllerBase
     [Authorize]
     public async Task<IActionResult> GetIndividualByCpf(string cpf)
     {
-        var response = await _personService.GetIndividualByCpfAsync(cpf);
-
-        if (!response.IsSuccess)
-            return StatusCode(response.StatusCode, response);
+        var response = await _personRouter.GetIndividualByCpfAsync(cpf);
 
         return StatusCode(response.StatusCode, response);
     }
@@ -72,10 +60,7 @@ public class PersonController : ControllerBase
     [HttpPost("legal")]
     public async Task<IActionResult> RegisterLegal([FromBody] RegisterLegalPersonRequestDto request)
     {
-        var response = await _personService.RegisterLegalAsync(request);
-
-        if (!response.IsSuccess)
-            return StatusCode(response.StatusCode, response);
+        var response = await _personRouter.RegisterLegalAsync(request);
 
         return StatusCode(response.StatusCode, response);
     }
@@ -84,10 +69,7 @@ public class PersonController : ControllerBase
     [Authorize]
     public async Task<IActionResult> UploadPhoto([FromForm] UploadPersonPhotoDto request)
     {
-        var response = await _personService.UploadPhotoAsync(request);
-
-        if (!response.IsSuccess)
-            return StatusCode(response.StatusCode, response);
+        var response = await _personRouter.UploadPhotoAsync(request);
 
         return StatusCode(response.StatusCode, response);
     }
@@ -96,10 +78,7 @@ public class PersonController : ControllerBase
     [Authorize]
     public async Task<IActionResult> UpdateLegal([FromBody] UpdateLegalPersonRequestDto request)
     {
-        var response = await _personService.UpdateLegalAsync(request);
-
-        if (!response.IsSuccess)
-            return StatusCode(response.StatusCode, response);
+        var response = await _personRouter.UpdateLegalAsync(request);
 
         return StatusCode(response.StatusCode, response);
     }
@@ -108,10 +87,7 @@ public class PersonController : ControllerBase
     [Authorize]
     public async Task<IActionResult> DeleteLegal([FromBody] DeleteLegalPersonDto request)
     {
-        var response = await _personService.DeleteLegalAsync(request);
-
-        if (!response.IsSuccess)
-            return StatusCode(response.StatusCode, response);
+        var response = await _personRouter.DeleteLegalAsync(request);
 
         return StatusCode(response.StatusCode, response);
     }
